% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_hmmprofile.R
\name{read_hmmprofile}
\alias{read_hmmprofile}
\title{Read and format a HMMprofile file from HMMbuild into tabular format}
\usage{
read_hmmprofile(file)
}
\arguments{
\item{file}{Filename}
}
\value{
data.frame
}
\description{
Read and format a HMMprofile file from HMMbuild into tabular format
}
\examples{
file <- system.file('extdata', 'example.hmmprofile.hmm', package='rhmmer')
read_hmmprofile(file)
}
