% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{read_tblout}
\alias{read_tblout}
\title{Read a file created through the '----tblout' option}
\usage{
read_tblout(file)
}
\arguments{
\item{file}{Filename}
}
\value{
data.frame
}
\description{
Read a file created through the '----tblout' option
}
\examples{
file <- system.file('extdata', 'example.tblout.txt', package='rhmmer')
read_tblout(file)
}
