% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{read_domtblout}
\alias{read_domtblout}
\title{Read a file created through the '----domtblout' option}
\usage{
read_domtblout(file)
}
\arguments{
\item{file}{Filename}
}
\value{
data.frame
}
\description{
Read a file created through the '----domtblout' option
}
\examples{
file <- system.file('extdata', 'example.domtblout.txt', package='rhmmer')
read_domtblout(file)
}
