% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhmmer.R
\docType{package}
\name{rhmmer}
\alias{rhmmer}
\alias{rhmmer-package}
\title{rhmmer: utilities for 'HMMER'}
\description{
Currently rhmmer exports exactly two functions:
}
\details{
\describe{
   \item{\code{read_tblout}}{Read a file made by HMMER '----tblout'}
   \item{\code{read_domtblout}}{Read a file made by HMMER '----domtblout'}
}
}
